refactor: shared inclusion logic, document matching in memory
