+++ packages/database/src/connectors/indexed-db.ts
@@ -12,9 +12,10 @@ import {
   TableData,
   TransactionDB,
   Schema,
-  Relation,
   constructSchema,
 } from '../types'
+import { validateDocuments, matchDocument } from '../helpers/memory'
+import { loadIncluded } from '../helpers/shared'
 
 const DB_NAME = 'zkopru'
 
@@ -93,69 +94,7 @@ export class IndexedDBConnector extends DB {
   ) {
     const table = this.schema[collection]
     if (!table) throw new Error(`Invalid collection: "${collection}"`)
-    const docs = [_doc].flat().map(doc => {
-      // insert defaults where needed
-      const defaults = {}
-      for (const key of Object.keys(table.rowsByName)) {
-        const row = table.rowsByName[key]
-        if (!row) throw new Error('Expected row to exist')
-        if (
-          row.default &&
-          (doc[row.name] === undefined || doc[row.name] === null)
-        ) {
-          Object.assign(defaults, {
-            [row.name]:
-              typeof row.default === 'function' ? row.default() : row.default,
-          })
-        }
-        const wipDoc = {
-          ...defaults,
-          ...doc,
-        }
-        if (
-          !row.optional &&
-          !row.relation &&
-          (wipDoc[row.name] === undefined || wipDoc[row.name] === null)
-        ) {
-          throw new Error(`NULL received for non-optional field "${row.name}"`)
-        }
-        if (
-          typeof wipDoc[row.name] !== 'undefined' &&
-          wipDoc[row.name] !== null
-        ) {
-          if (row.type === 'Bool' && typeof wipDoc[row.name] !== 'boolean') {
-            throw new Error(
-              `Unrecognized value ${wipDoc[row.name]} for type Bool`,
-            )
-          } else if (
-            row.type === 'Int' &&
-            typeof wipDoc[row.name] !== 'number'
-          ) {
-            throw new Error(
-              `Unrecognized value ${wipDoc[row.name]} for type Int`,
-            )
-          } else if (
-            row.type === 'String' &&
-            typeof wipDoc[row.name] !== 'string'
-          ) {
-            throw new Error(
-              `Unrecognized value ${wipDoc[row.name]} for type String`,
-            )
-          } else if (
-            row.type === 'Object' &&
-            typeof wipDoc[row.name] !== 'object'
-          ) {
-            throw new Error(
-              `Unrecognized value ${wipDoc[row.name]} for type Object`,
-            )
-          }
-        }
-      }
-      return {
-        ...defaults,
-        ...doc,
-      }
-    })
+    const docs = validateDocuments(table, _doc)
     if (!this.db) throw new Error('DB is not initialized')
     const tx = _tx || this.db.transaction(collection, 'readwrite')
     const createPromises = docs.map(doc => {
@@ -174,57 +113,6 @@ export class IndexedDBConnector extends DB {
       limit: 1,
     })
     return obj === undefined ? null : obj
-  }
-
-  async loadIncluded(
-    collection: string,
-    options: { models: any[]; include?: any },
-  ) {
-    const { models, include } = options
-    if (!include) return
-    const table = this.schema[collection]
-    if (!table) throw new Error(`Unable to find table ${collection} in schema`)
-    for (const key of Object.keys(include)) {
-      const relation = table.relations[key]
-      if (!relation) {
-        throw new Error(`Unable to find relation ${key} in ${collection}`)
-      }
-      if (include[key]) {
-        await this.loadIncludedModels(
-          models,
-          relation,
-          typeof include[key] === 'object' ? include[key] : undefined,
-        )
-      }
-    }
-  }
-
-  private async loadIncludedModels(
-    models: any[],
-    relation: Relation & { name: string },
-    include?: any,
-  ) {
-    const values = models.map(model => model[relation.localField])
-    // load relevant submodels
-    const submodels = await this._findMany(relation.foreignTable, {
-      where: {
-        [relation.foreignField]: values,
-      },
-      include: include as any, // load subrelations if needed
-    })
-    // key the submodels by their relation field
-    const keyedSubmodels = {}
-    for (const submodel of submodels) {
-      // assign to the models
-      keyedSubmodels[submodel[relation.foreignField]] = submodel
-    }
-    // Assign submodel onto model
-    for (const model of models) {
-      const submodel = keyedSubmodels[model[relation.localField]]
-      Object.assign(model, {
-        [relation.name]: submodel || null,
-      })
-    }
   }
 
   async findMany(collection: string, options: FindManyOptions) {
@@ -299,9 +187,11 @@ export class IndexedDBConnector extends DB {
       const query = index.keys.map(k => options.where[k])
       const result = await txIndex.getAll(query)
       const found = result.filter(i => !!i)
-      await this.loadIncluded(collection, {
+      await loadIncluded(collection, {
         models: found,
         include: options.include,
+        findMany: this._findMany.bind(this),
+        table,
       })
       return found
     }
@@ -339,70 +229,20 @@ export class IndexedDBConnector extends DB {
     } else {
       cursor = await tx.objectStore(collection).openCursor()
     }
-    const matchDoc = (where: WhereClause, doc: any) => {
-      for (const [key, val] of Object.entries(where)) {
-        if (typeof val === 'object' && !Array.isArray(val) && val !== null) {
-          if (typeof val.ne !== 'undefined' && doc[key] === val.ne) {
-            return false
-          }
-          if (typeof val.lt !== 'undefined' && doc[key] >= val.lt) {
-            return false
-          }
-          if (typeof val.lte !== 'undefined' && doc[key] > val.lte) {
-            return false
-          }
-          if (typeof val.gt !== 'undefined' && doc[key] <= val.gt) {
-            return false
-          }
-          if (typeof val.gte !== 'undefined' && doc[key] < val.gte) {
-            return false
-          }
-        } else if (Array.isArray(val)) {
-          let exists = false
-          for (const v of val) {
-            if (v === null && typeof doc[key] === 'undefined') {
-              exists = true
-              break
-            }
-            if (doc[key] === v) {
-              exists = true
-              break
-            }
-          }
-          if (!exists) return false
-        } else if (
-          val === null &&
-          typeof doc[key] !== 'undefined' &&
-          doc[key] !== null
-        ) {
-          return false
-        } else if (val !== null && doc[key] !== val) {
-          return false
-        }
-      }
-      return true
-    }
     const { where, limit } = options
     while (cursor) {
       if (typeof limit === 'number' && found.length >= limit) break
       const obj = cursor.value
-      const topWhere = { ...where, OR: undefined }
-      const or = where.OR || []
-      const matched = matchDoc(topWhere, obj)
-      if (or.length === 0 && matched) {
+      if (matchDocument(where, obj)) {
         found.push(obj)
-      }
-      for (const _where of or) {
-        if (matchDoc(_where, obj) && matched) {
-          found.push(obj)
-          break
-        }
       }
       cursor = await cursor.continue()
     }
-    await this.loadIncluded(collection, {
+    await loadIncluded(collection, {
       models: found,
       include: options.include,
+      findMany: this._findMany.bind(this),
+      table,
     })
     return found
   }
