+++ packages/database/src/connectors/postgres.ts
@@ -12,7 +12,6 @@ import {
   UpsertOptions,
   DeleteManyOptions,
   TableData,
-  Relation,
   TransactionDB,
 } from '../types'
 import {
@@ -24,6 +23,7 @@ import {
   countSql,
   upsertSql,
 } from '../helpers/sql'
+import { loadIncluded } from '../helpers/shared'
 
 export class PostgresConnector extends DB {
   db: Client
@@ -94,59 +94,6 @@ export class PostgresConnector extends DB {
     return obj === undefined ? null : obj
   }
 
-  // load related models
-  async loadIncluded(
-    collection: string,
-    options: { models: any[]; include?: any },
-  ) {
-    const { models, include } = options
-    if (!include) return
-    const table = this.schema[collection]
-    if (!table) throw new Error(`Unable to find table ${collection} in schema`)
-    for (const key of Object.keys(include)) {
-      // for each relation to include
-      const relation = table.relations[key]
-      if (!relation)
-        throw new Error(`Unable to find relation ${key} in ${collection}`)
-      if (include[key]) {
-        await this.loadIncludedModels(
-          models,
-          relation,
-          typeof include[key] === 'object' ? include[key] : undefined,
-        )
-      }
-    }
-  }
-
-  // load and assign submodels, mutates the models array supplied
-  private async loadIncludedModels(
-    models: any[],
-    relation: Relation & { name: string },
-    include?: any,
-  ) {
-    const values = models.map(model => model[relation.localField])
-    // load relevant submodels
-    const submodels = await this._findMany(relation.foreignTable, {
-      where: {
-        [relation.foreignField]: values,
-      },
-      include: include as any, // load subrelations if needed
-    })
-    // key the submodels by their relation field
-    const keyedSubmodels = {}
-    for (const submodel of submodels) {
-      // assign to the models
-      keyedSubmodels[submodel[relation.foreignField]] = submodel
-    }
-    // Assign submodel onto model
-    for (const model of models) {
-      const submodel = keyedSubmodels[model[relation.localField]]
-      Object.assign(model, {
-        [relation.name]: submodel || null,
-      })
-    }
-  }
-
   async findMany(collection: string, options: FindManyOptions) {
     return this.lock.acquire('read', async () =>
       this._findMany(collection, options),
@@ -176,9 +123,11 @@ export class PostgresConnector extends DB {
       }
     }
     const { include } = options
-    await this.loadIncluded(collection, {
+    await loadIncluded(collection, {
       models: rows,
       include,
+      findMany: this._findMany.bind(this),
+      table,
     })
     return rows
   }
